BE#18

+++ app/services/service_company.py
@@ -31,9 +31,6 @@ class Service_company:
 
 
     async def create_company(self, payload:CreateCompany, owner_id: int) -> CompanySchema:
-        # error if no name
-        if payload.company_name == "":
-            raise HTTPException(status_code=status.HTTP_422_UNPROCESSABLE_ENTITY, detail="No name provided")
         # create company
         query = insert(Company).values(
             company_name = payload.company_name,
@@ -70,6 +67,9 @@ class Service_company:
     async def delete_member(self, member_id: int) -> status:
         # check if user is owner
         await self.is_owner()
+        # check if user doesn't try to delete himself
+        if self.user.user_id == member_id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You cannot delete yourself")
         #delete member
         await Service_membership(db=self.db, company_id=self.company_id).delete_membership(member_id=member_id)
         return status.HTTP_200_OK
@@ -85,8 +85,11 @@ class Service_company:
 
     async def leave_company(self) -> status:
         # check if user is member
-        if not await self.is_member():
-            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="You are not a member of this company")
+        if not await self.is_member(member_id=self.user.user_id):
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You are not a member of this company")
+        # check if user is owner
+        if await Service_membership(db=self.db, company_id=self.company_id).get_members(member_id=self.user.user_id, role="owner"):
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You must give ownership of the company to admin before leaving")
         # leave company
         await Service_membership(db=self.db, company_id=self.company_id).delete_membership(member_id=self.user.user_id)
         return status.HTTP_200_OK
@@ -97,9 +100,9 @@ class Service_company:
         await self.is_owner()
         # check if user exists
         await Service_user(db=self.db, user=self.user).get_by_id(user_id=member_id)
-        # check if user is member
-        if not await self.is_member():
-            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="User is not a member of this company")
+        # check if user exists and is member
+        if not await self.is_member(member_id=member_id):
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="User is not a member of this company")
         # add admin to members of the company
         membership = await Service_membership(db=self.db, company_id=self.company_id).change_membership(member_id=member_id, role="admin")
         return membership
@@ -110,42 +113,62 @@ class Service_company:
         await self.is_owner()
         # check if user exists
         await Service_user(db=self.db, user=self.user).get_by_id(user_id=member_id)
-        # check if user is member
-        if not await self.is_member():
-            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="User is not a member of this company")
+        # check if user exists and is member
+        if not await self.is_member(member_id=member_id):
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="User is not a member of this company")
         # downgrade member
         membership = await Service_membership(db=self.db, company_id=self.company_id).change_membership(member_id=member_id, role=new_role)
         return membership
+    
+
+    async def give_ownership(self, member_id: int) -> MembershipsList:
+        # check if user's company
+        await self.is_owner()
+        # check if user_to_be_owner exists
+        await Service_user(db=self.db, user=self.user).get_by_id(user_id=member_id)
+        # check if user_to_be_owner is admin
+        if not await Service_company(db=self.db, company_id=self.company_id).is_admin_owner(member_id=member_id):
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="User is not admin")
+        # downgrade owner
+        new_admin = await Service_membership(db=self.db, company_id=self.company_id).change_membership(member_id=self.user.user_id, role="admin")
+        # upgrade admin 
+        new_owner = await Service_membership(db=self.db, company_id=self.company_id).change_membership(member_id=member_id, role="owner")
+        query = update(Company).where(Company.company_id == self.company_id).values(
+            company_owner_id = member_id
+        ).returning(Company)
+        await self.db.execute(query)
+        return MembershipsList(users=[new_owner, new_admin])
         
 
     def get_changed_values(self, payload:CompanyUpdate) -> CompanyUpdate:
         changed_values = {x[0]:x[1] for x in payload if x[1]}
         # if nothing changed
         if changed_values == {}:
-            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Nothing to change")
+            raise HTTPException(status_code=status.HTTP_422_UNPROCESSABLE_ENTITY, detail="Nothing to change")
         return changed_values
 
     async def is_owner(self) -> status:
         #check if company exists
         await self.get_by_id(company_id=self.company_id)
         # check if user is owner
-        company_members = await Service_membership(db=self.db, company_id=self.company_id).get_members(role="owner")
-        if self.user.user_id not in [member.membership_user_id for member in company_members.users]:
+        member = await Service_membership(db=self.db, company_id=self.company_id).get_members(role="owner", member_id=self.user.user_id)
+        if not bool(member):
             raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="It's not your company")
         return status.HTTP_200_OK
     
-    async def is_admin_owner(self) -> status:
+    async def is_admin_owner(self, member_id:int) -> bool:
         # check if company exists
         await self.get_by_id(company_id=self.company_id)
         # check if user is admin or owner
-        company_members = await Service_membership(db=self.db, company_id=self.company_id).get_members(role=["admin","owner"])
-        if self.user.user_id not in [member.membership_user_id for member in company_members.users]:
-            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You don't have permission")
-        return status.HTTP_200_OK
+        member = await Service_membership(db=self.db, company_id=self.company_id).get_members(role=["admin","owner"], member_id=member_id)
+        return bool(member)
     
 
-    async def is_member(self) -> bool:
+    async def is_member(self, member_id:int) -> bool:
         # check if company exists
         await self.get_by_id(company_id=self.company_id)
-        company_members = await Service_membership(db=self.db, company_id=self.company_id).get_members()
-        return self.user.user_id in [member.membership_user_id for member in company_members.users]
+        # check if member exists
+        user = await Service_user(db=self.db, user=self.user).get_by_id(user_id=member_id)
+        await Service_user(db=self.db, user=user).get_by_id(user_id=member_id)
+        member = await Service_membership(db=self.db, company_id=self.company_id).get_members(member_id=member_id)
+        return bool(member)
